requirements.txt
# Core dependencies
python-telegram-bot==20.7
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0

# Google Sheets integration
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
gspread==5.12.3

# Data validation and models
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM integration (OpenAI for NLP parsing)
openai==1.10.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

